New GDNative API
This uses the new GDNative API from Godot 4. Major change is in the
ability to wrap the builtin types as well, so we don't have to
implement wrappers manually anymore (not on Godot side nor here).

The bindings generator now takes that into account to create the new
wrappers, including a case for utility functions, which are also exposed
from Godot (like print() which now should be called from Utilities).

A few special functions are still provided for convenience, like
print_error().

Some builtin types that are used in math are reimplemented here with a
copy from Godot source (with a few compatibility changes). This is so
you can make use of inlined functions. It is possible that this might
cause problems when marshalling from or to the engine, given difference
in memory alignment between compilers. I haven't experienced this and
that was pretty much already the case before, so it shouldn't be a
problem.