Fix CMake generation on Windows
On Windows, paths are mixed by os.path.join with different slash directions. This should generally be fine but when using the print function we print "C:\\mypath/myfile.cpp" as "C:\mypath/myfile.cpp" which is unescaped on the backward slash. To fix this I propose to replace "\\" with "/" just before printing it out for CMake usage. I would suggest using pathlib but it is not included by default on python versions <= 3.3 which may still be an issue.