Make SCons decide if bindings need to be generated
Adds the binding generator as a SCons builder,
which allows using the api.json file as a dependency.

SCons will automatically trigger the binding generator
during build if a generated file is missing, or if the api.json changes.

This allows better utilization of the build cache,
as the bindings (and thus the library)
are not regenerated every time SCons runs.

Both arguments `auto` and `yes` for `generate_bindings` are accepted,
binding generation is only entirely skipped when passing `no`.