Remove godot-headers submodule, copy files directly
With the new GDExtension API, the headers are just two files, one of
which is generated and needs to always be kept in sync with the Godot
engine version.

So there's little practical use for using godot-headers as a submodule
anymore, and it only makes godot-cpp updates more cumbersome.

Custom headers (i.e. a custom API JSON) can still be used by passing
the `headers_dir` SCons option.