Add DISABLE_OBJECT_BINDING_GUARD compiler flag
The guard in `Wrapped::Wrapped(const StringName)` detects the partial
initialization of `godot::Object` subclasses.  This is needed because
using `new` to allocate a `godot::Object` subclass will result in
unexpected crashes because the bindings for the instance have not been
set.

The side effect of this guard is that it also prevents any use of a
`godot::Object` subclass on the stack.

There are godot native types such as `godot::RegEx` that are safe to use
on the stack, and are used that way throughout the godot source.
Similarly, custom classes that subclass `godot::Object` may also be used
safely on the stack[^1].

In this commit, I add a compiler flag to disable this guard.  This will
allow godot-cpp users who understand the safety implications to use
`godot::Object` subclasses on the stack.

[^1]: https://github.com/godotengine/godot-cpp/pull/1446#issuecomment-2351331480